处理冲突
