+++ lib/coolq.js
@@ -1,6 +1,7 @@
 /**
  * 酷Q用公共函数
  */
+const { isInGoogle, gcj_wgs_bored, gcj_bd, coordsRound } = require('ishisashimap');
 
 const faces = {
     0: '惊讶', 1: '撇嘴', 2: '色', 3: '发呆', 4: '得意', 5: '流泪', 6: '害羞', 7: '闭嘴', 8: '睡', 9: '大哭',
@@ -33,17 +34,19 @@ const faces = {
 /**
  * 去除接收訊息中的 CQ 碼（酷 Q 專用碼，包括表情、繪文字、相片等資料），將其換為「[表情名稱]」、「[图片]」等文字
  * @param  {string} Message 已解碼並轉為 UTF-8 之後的訊息
- * @return {object} 去除 CQ 碼之後的文字
+ * @return {string} 去除 CQ 碼之後的文字
  */
 const parseMessage = (message) => {
     let images = [];
     let records = [];
     let at = [];
+    let coords = {};
 
     let text = message.replace(/\n/gu, '&#10;').replace(/\[CQ:([^,]*?)\]/gu, '[CQ:$1,]').replace(/\[CQ:(.*?)((?:,).*?)\]/gu, (_, type, param) => {
         let tmp;
         let tmp1;
         let tmp2;
+        let tmp3;
         switch (type) {
             case 'face':
                 // [CQ:face,id=13]
@@ -86,10 +89,15 @@ const parseMessage = (message) => {
             case 'rich':
                 // 分享音樂 [CQ:rich,url=http://music.163.com/song/504733843/?userid=263400453,text= 新宝島 BENI ]
                 // 協議不支援 [CQ:rich,text=QQ天气提示 升级QQ查看天气]
-                tmp1 = param.match(/(?:^|,)url=(.*?)(?:,|$)/u);
-                tmp2 = param.match(/(?:^|,)text=(.*?)(?:,|$)/u);
-                if (tmp1 && tmp1[1]) {
-                    return `[分享链接：${tmp1[1]}]`;
+                tmp1 = param.match(/(?:^|,)text=(.*?)(?:,|$)/u);
+                tmp2 = param.match(/(?:^|,)url=(.*?)(?:,|$)/u);
+                if (tmp2 && tmp2[1]) {
+                    tmp = ['[分享]'];
+                    if (tmp1 && tmp1[1]) {
+                        tmp.push(tmp1[1]);
+                    }
+                    tmp.push(tmp2[1]);
+                    return tmp.join('\n');
                 } else if (tmp2 && tmp2[1]) {
                     return tmp2[1];
                 } else {
@@ -124,47 +132,70 @@ const parseMessage = (message) => {
 
             case 'share':
                 // [CQ:share,url=http://www.bilibili.com/video/av42585280?share_medium=android&amp;share_source=qq&amp;bbid=XZ97F38904CBFC1747BFE02321AFCB3A3D933&amp;ts=1549692084426,title=三天之内,content=给生活找点乐子~,image=http://url.cn/5AEq2ju]
-                tmp = param.match(/(?:^|,)url=(.*?)(?:,|$)/u);
-                if (tmp && tmp[1]) {
-                    return `[分享链接：${tmp[1]}]`;
-                } else {
-                    return '[分享]';
-                }
-
-            case 'hb':
-                // [CQ:hb,title=恭喜发财]
-                tmp = param.match(/(?:^|,)title=(.*?)(?:,|$)/u);
-                if (tmp && tmp[1]) {
-                    return `[红包：${tmp[1]}]`;
-                } else {
-                    return '[红包]';
-                }
-
-            case 'sign':
-                // [CQ:sign,title=我过来签个到啦,image=https://p.qpic.cn/qunsign/0/sign_30cc0f793397325b74be99fabd5f9cfcjs0qje77/750]
-                tmp = param.match(/(?:^|,)title=(.*?)(?:,|$)/u);
-                if (tmp && tmp[1]) {
-                    return `[签到：${tmp[1]}]`;
-                } else {
-                    return '[签到]';
-                }
-
-            case 'location':
-                // [CQ:location,lat=23.081961,lon=113.453941,title=大吉沙,content=广东省广州市黄埔区,style=1]
-                tmp = [];
+                tmp = ['[分享]'];
                 tmp1 = param.match(/(?:^|,)title=(.*?)(?:,|$)/u);
                 tmp2 = param.match(/(?:^|,)content=(.*?)(?:,|$)/u);
+                tmp3 = param.match(/(?:^|,)url=(.*?)(?:,|$)/u);
                 if (tmp1 && tmp1[1]) {
                     tmp.push(tmp1[1]);
                 }
                 if (tmp2 && tmp2[1]) {
                     tmp.push(tmp2[1]);
                 }
-                if (tmp.length > 0) {
-                    return `[位置：${tmp.join(' ')}]`;
-                } else {
-                    return '[位置]';
+                if (tmp3 && tmp3[1]) {
+                    tmp.push(tmp3[1]);
                 }
+                return tmp.join('\n');
+
+            case 'hb':
+                // [CQ:hb,title=恭喜发财]
+                tmp = ['[红包]'];
+                tmp1 = param.match(/(?:^|,)title=(.*?)(?:,|$)/u);
+                if (tmp1 && tmp1[1]) {
+                    tmp.push(tmp1[1]);
+                }
+                return tmp.join('\n');
+
+            case 'sign':
+                // [CQ:sign,title=我过来签个到啦,image=https://p.qpic.cn/qunsign/0/sign_30cc0f793397325b74be99fabd5f9cfcjs0qje77/750]
+                tmp = ['[签到]'];
+                tmp1 = param.match(/(?:^|,)title=(.*?)(?:,|$)/u);
+                if (tmp1 && tmp1[1]) {
+                    tmp.push(tmp1[1]);
+                }
+                return tmp.join('\n');
+
+            case 'location':
+                // [CQ:location,lat=23.081961,lon=113.453941,title=大吉沙,content=广东省广州市黄埔区,style=1]
+                tmp = ['[位置]'];
+                tmp1 = param.match(/(?:^|,)title=(.*?)(?:,|$)/u);
+                tmp2 = param.match(/(?:^|,)content=(.*?)(?:,|$)/u);
+                tmp3 = param.match(/(?:^|,)lat=(.*?)(?:,|$)/u);
+                let tmp4 = param.match(/(?:^|,)lon=(.*?)(?:,|$)/u);
+                if (tmp1 && tmp1[1]) {
+                    tmp.push(tmp1[1]);
+                }
+                if (tmp2 && tmp2[1]) {
+                    tmp.push(tmp2[1]);
+                }
+                if (tmp3 && tmp3[1] && tmp4 && tmp4[1]) {
+                    let incoords = { lat: Number(tmp3[1]), lon: Number(tmp4[1]) };
+                    if (isInGoogle(incoords)) {
+                        let wgs = gcj_wgs_bored(incoords, false);
+                        let bd = gcj_bd(incoords, false);
+                        coords.wgs = wgs;
+                        coords.gcj = incoords;
+                        coords.bd = bd;
+                        // 非原始數據比原始數據多保留一位
+                        wgs = coordsRound(wgs, 7);
+                        bd = coordsRound(bd, 7);
+                        tmp.push(`WGS-84: ${wgs.lat},${wgs.lon}`, `GCJ-02: ${incoords.lat},${incoords.lon}`, `BD-09: ${bd.lat},${bd.lon}`);
+                    } else {
+                        coords.wgs = incoords;
+                        tmp.push(`WGS-84: ${incoords.lat},${incoords.lon}`);
+                    }
+                }
+                return tmp.join('\n');
 
             case 'contact':
                 // 群邀請 [CQ:contact,id=609486016,type=group]
@@ -172,9 +203,17 @@ const parseMessage = (message) => {
                 tmp1 = param.match(/(?:^|,)id=(.*?)(?:,|$)/u);
                 tmp2 = param.match(/(?:^|,)type=(.*?)(?:,|$)/u);
                 if (tmp1 && tmp1[1] && tmp2 && tmp2[1] === 'group') {
-                    return `[群邀请：${tmp1[1]}]`;
+                    tmp = ['[群邀请]'];
+                    if (tmp1 && tmp1[1]) {
+                        tmp.push(tmp1[1]);
+                    }
+                    return tmp.join('\n');
                 } else if (tmp1 && tmp1[1] && tmp2 && tmp2[1] === 'qq') {
-                    return `[好友邀请：${tmp1[1]}]`;
+                    tmp = ['[好友邀请]'];
+                    if (tmp1 && tmp1[1]) {
+                        tmp.push(tmp1[1]);
+                    }
+                    return tmp.join('\n');
                 } else {
                     return '[邀请]';
                 }
@@ -184,42 +223,50 @@ const parseMessage = (message) => {
                 // 網易雲 [CQ:music,type=163,id=509842]
                 // QQ 音樂 [CQ:music,type=qq,id=200732275]
                 // 蝦米 [CQ:music,type=xiami,id=1769370187]
-                tmp = param.match(/(?:^|,)type=(.*?)(?:,|$)/u);
-                tmp1 = param.match(/(?:^|,)url=(.*?)(?:,|$)/u);
-                tmp2 = param.match(/(?:^|,)id=(.*?)(?:,|$)/u);
-                if (tmp && tmp[1] === 'custom' && tmp1 && tmp1[1]) {
-                    return `[分享音乐：${tmp1[1]}]`;
-                } else if (tmp && tmp[1] === '163' && tmp2 && tmp2[1]) {
-                    return `[分享音乐：https://music.163.com/#/song?id=${tmp2[1]}]`
-                } else if (tmp && tmp[1] === 'qq' && tmp2 && tmp2[1]) {
-                    return `[分享音乐：https://y.qq.com/n/yqq/song/${tmp2[1]}_num.html]`
-                } else if (tmp && tmp[1] === 'xiami' && tmp2 && tmp2[1]) {
-                    return `[分享音乐：https://www.xiami.com/song/${tmp2[1]}]`
-                } else {
-                    return '[分享音乐]';
+                tmp = ['[分享音乐]'];
+                tmp1 = param.match(/(?:^|,)type=(.*?)(?:,|$)/u);
+                tmp2 = param.match(/(?:^|,)url=(.*?)(?:,|$)/u);
+                tmp3 = param.match(/(?:^|,)id=(.*?)(?:,|$)/u);
+                if (tmp1 && tmp1[1] === 'custom') {
+                    if (tmp2 && tmp2[1]) {
+                        tmp.push(tmp2[1]);
+                    }
+                } else if (tmp1 && tmp1[1] === '163') {
+                    if (tmp3 && tmp3[1]) {
+                        tmp.push(`https://music.163.com/#/song?id=${tmp3[1]}`);
+                    }
+                } else if (tmp1 && tmp1[1] === 'qq') {
+                    if (tmp3 && tmp3[1]) {
+                        tmp.push(`https://y.qq.com/n/yqq/song/${tmp3[1]}_num.html`);
+                    }
+                } else if (tmp1 && tmp1[1] === 'xiami') {
+                    if (tmp3 && tmp3[1]) {
+                        tmp.push(`https://www.xiami.com/song/${tmp3[1]}`);
+                    }
                 }
+                return tmp.join('\n');
 
             case 'rps':
                 // [CQ:rps,type=1]
-                tmp = param.match(/(?:^|,)type=(.*?)(?:,|$)/u);
-                if (tmp && tmp[1] === '1') {
-                    return '[猜拳：石头]';
-                } else if (tmp && tmp[1] === '2') {
-                    return '[猜拳：剪刀]';
-                } else if (tmp && tmp[1] === '3') {
-                    return '[猜拳：布]';
-                } else {
-                    return '[猜拳]';
+                tmp = ['[猜拳]'];
+                tmp1 = param.match(/(?:^|,)type=(.*?)(?:,|$)/u);
+                if (tmp1 && tmp1[1] === '1') {
+                    tmp.push('石头');
+                } else if (tmp1 && tmp1[1] === '2') {
+                    tmp.push('剪刀');
+                } else if (tmp1 && tmp1[1] === '3') {
+                    tmp.push('布');
                 }
+                return tmp.join('\n');
 
             case 'dice':
                 // [CQ:dice,type=1]
-                tmp = param.match(/(?:^|,)type=(.*?)(?:,|$)/u);
-                if (tmp && tmp[1]) {
-                    return `[骰子：${tmp[1]}]`;
-                } else {
-                    return '[骰子]';
+                tmp = ['[骰子]'];
+                tmp1 = param.match(/(?:^|,)type=(.*?)(?:,|$)/u);
+                if (tmp1 && tmp1[1]) {
+                    tmp.push(tmp1[1]);
                 }
+                return tmp.join('\n');
 
             case 'shake':
                 // [CQ:shake]
@@ -242,6 +289,7 @@ const parseMessage = (message) => {
             images: images,
             records: records,
             ats: ats,
+            coords: coords,
         },
         raw: message,
     };
